refactor(catch-spotify-error): move to `@lib/utils`
