+++ apps/statistics/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './format-artists'
 export * from './format-tracks'
-export * from './catch-spotify-error'
