+++ libs/utils/src/catch-spotify-error.spec.ts
